refactor: use id as path parameter in product
