+++ .changeset/smooth-eyes-remain.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+refactor: use id as path parameter to manage products
